Move `pkg/beacon/state` out of `pkg/beacon`
This change will allow reusing the state
machine by other applications.
